Reworked auth guard strategy for GoogleDriveContext
- Implemented its error-handling logic in ImageUpload and ImageList components
- Added AbortControlers to the Axios calls where necessary

+++ src/hooks/GoogleDriveContext.tsx
@@ -1,7 +1,8 @@
 import type { TokenResponse } from "@react-oauth/google";
 import type { AxiosResponse } from "axios";
 import type { PropsWithChildren } from "react";
-import { useGoogleLogin } from "@react-oauth/google";
+import { useEffect } from "react";
+import { useGoogleLogin, hasGrantedAnyScopeGoogle } from "@react-oauth/google";
 import { useState, createContext, useContext } from "react";
 import axios from "axios";
 
@@ -28,6 +29,7 @@ type TokenResponseError = Pick<
 type TokenInfo = {
   tokenExpiration: Date;
 };
+
 type FileUploadJSONResponse = (FileUploadSuccess & GoogleDriveError) | false;
 type FileDownloadJSONResponse = ArrayBuffer | GoogleDriveError | false;
 type FileDeletedJSONResponse = GoogleDriveError | "";
@@ -36,10 +38,11 @@ type FilesListJSONResponse = {
 } & GoogleDriveError;
 
 type GoogleDriveContextType = {
-  uploadFile({
-    file,
-    metadata,
-  }: FileParams): Promise<AxiosResponse<FileUploadJSONResponse, unknown>>;
+  hasScope: boolean;
+  uploadFile(
+    { file, metadata }: FileParams,
+    signal?: AbortSignal
+  ): Promise<AxiosResponse<FileUploadJSONResponse, unknown>>;
   fetchList(
     signal?: AbortSignal
   ): Promise<AxiosResponse<FilesListJSONResponse, unknown>>;
@@ -54,6 +57,7 @@ type GoogleDriveContextType = {
 };
 
 const googleDriveContext = createContext<GoogleDriveContextType>({
+  hasScope: false,
   uploadFile: () => {
     throw new Error("Google Drive context is uninitialized");
   },
@@ -78,11 +82,11 @@ const DISCOVERY_DOC =
 export function GoogleDriveProvider({ children }: PropsWithChildren) {
   const { user } = useUser();
 
+  const [isLoaded, setIsLoaded] = useState(false);
+  const [hasScope, setHasScope] = useState(false);
   const [userTokens, setUserTokens] = useState<
     TokenResponseSuccess & TokenInfo
   >();
-
-  const [isLoaded, setIsLoaded] = useState(false);
 
   async function initGapiClient() {
     try {
@@ -124,23 +128,31 @@ export function GoogleDriveProvider({ children }: PropsWithChildren) {
     scope,
   });
 
-  async function guardTokens() {
-    const promise = new Promise<string>((resolve, reject) => {
-      if (!isLoaded) reject("Drive operation: Client is not ready");
-      if (!user) {
-        setUserTokens(undefined);
-        reject("User not logged in");
-      }
+  function authGuard(): "OK" | "Authorizing" | "Unauthorized" {
+    if (!isLoaded)
+      throw new Error("Unauthorized", { cause: "Google Drive is not loaded" });
+    if (!user) {
+      setUserTokens(undefined);
+      throw new Error("User not authenticated");
+    }
 
-      if (userTokens === undefined) requestAccess();
-      else if (userTokens.tokenExpiration > new Date()) resolve("OK");
-    });
-
-    return promise;
+    if (userTokens === undefined) {
+      requestAccess({ prompt: "" });
+      return "Authorizing";
+    } else if (userTokens.tokenExpiration > new Date()) {
+      return "OK";
+    } else {
+      setUserTokens(undefined);
+      throw new Error("Unauthorized", { cause: "Session expired" });
+    }
   }
 
-  const uploadFile = async ({ file, metadata }: FileParams) => {
-    await guardTokens();
+  const uploadFile = async (
+    { file, metadata }: FileParams,
+    signal: AbortSignal
+  ) => {
+    const authStatus = authGuard();
+    if (authStatus !== "OK") return Promise.reject(authStatus);
 
     metadata.parents = [spaces];
 
@@ -159,6 +171,7 @@ export function GoogleDriveProvider({ children }: PropsWithChildren) {
         headers: {
           Authorization: `Bearer ${userTokens?.access_token}`,
         },
+        signal,
       }
     );
 
@@ -166,7 +179,8 @@ export function GoogleDriveProvider({ children }: PropsWithChildren) {
   };
 
   const fetchList = async (signal?: AbortSignal) => {
-    await guardTokens();
+    const authStatus = authGuard();
+    if (authStatus !== "OK") return Promise.reject(authStatus);
 
     const request = axios.get("https://www.googleapis.com/drive/v3/files", {
       params: {
@@ -183,7 +197,8 @@ export function GoogleDriveProvider({ children }: PropsWithChildren) {
   };
 
   const fetchFile = async (file: gapi.client.drive.File) => {
-    await guardTokens();
+    const authStatus = authGuard();
+    if (authStatus !== "OK") return Promise.reject(authStatus);
 
     const request = axios.get(
       `https://www.googleapis.com/drive/v3/files/${file.id}`,
@@ -200,7 +215,8 @@ export function GoogleDriveProvider({ children }: PropsWithChildren) {
   };
 
   const deleteFile = async (file: gapi.client.drive.File) => {
-    await guardTokens();
+    const authStatus = authGuard();
+    if (authStatus !== "OK") return Promise.reject(authStatus);
 
     const request = axios.delete(
       `https://www.googleapis.com/drive/v3/files/${file.id}`,
@@ -214,9 +230,15 @@ export function GoogleDriveProvider({ children }: PropsWithChildren) {
     return request;
   };
 
+  useEffect(() => {
+    if (!isLoaded || !userTokens) return;
+    setHasScope(hasGrantedAnyScopeGoogle(userTokens, scope));
+  }, [userTokens]);
+
   return (
     <googleDriveContext.Provider
       value={{
+        hasScope,
         uploadFile,
         fetchList,
         fetchFile,
